Raw data parsing script
Initial script to associate scraped data with nutrition data.

So far most of my effort has been in splitting the data because it contained both portions in parenthesis and without, as well as embedded commas that had to be maintained. Ended up using the "re" library to split based on multiple criteria. Next effort will be towards making dataframes of nutritional data for each recipe.